Moved custom validator to shared folder
