+++ src/app/admin/user-add/user-add.component.html
@@ -92,3 +92,4 @@
   <!-- {{ userRegisterationForm.value | json }} -->
 </pre>
 </div>
+<!-- TODO add eye like icon to show password  -->
